Close blocked threads
Here's the deal, the workers are all waiting on urls concurrently.  Then
sends those urls back to be processed synchronously.  After processing,
another URL *can* be sent to the worker.

In effect, we had a deadlock loop.  We would leave those workers alive
if we returned out of the function, just waiting for more URLs that
would never come.  I needed to find some way to pump that loop and exit
the goroutines.

Sending `nil` as a sigil value allows us to return out of the goroutine
and avoid the deadlock condition of sending more data to the receiver
just to loop back and wait for another URL to be present.